# docs/openapi.yaml
openapi: 3.1.0
info:
  title: WikiMind Gateway
  description: Local LLM Knowledge OS — Personal API
  version: 0.1.0
paths:
...
  /api/settings:
    get:
      tags:
      - Settings
      summary: Get All Settings
      description: Return all application settings, with DB overrides applied.
      operationId: get_all_settings_api_settings_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllSettingsResponse'
    patch:
      tags:
      - Settings
      summary: Update Settings
      description: Update runtime settings. Changes persist to DB across restarts.
      operationId: update_settings_api_settings_patch
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SettingsUpdateRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettingsUpdateResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /auth/login/{provider}:
    get:
      tags:
      - Auth
      summary: Login
      description: Redirect to OAuth2 provider's authorize URL.
      operationId: login_auth_login__provider__get
      parameters:
      - name: provider
        in: path
        required: true
        schema:
          type: string
          title: Provider
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /auth/callback:
    get:
      tags:
      - Auth
      summary: Auth Callback
      description: Handle OAuth2 callback — exchange code for token, upsert user,
        set cookie.
      operationId: auth_callback_auth_callback_get
      parameters:
      - name: code
        in: query
        required: true
        schema:
          type: string
          title: Code
      - name: state
        in: query
        required: true
        schema:
          type: string
          title: State
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /auth/magic-link:
    post:
      tags:
      - Auth
      summary: Request Magic Link
      description: 'Request a magic link for passwordless email login.


        Generates an HMAC-signed token encoding the email and timestamp.

        In non-production mode, the token is returned in the response for

        CLI/testing use. In production, an email would be sent (not yet

        implemented).'
      operationId: request_magic_link_auth_magic_link_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MagicLinkRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MagicLinkResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /auth/magic-link/verify:
    post:
      tags:
      - Auth
      summary: Verify Magic Link
      description: 'Verify a magic link token and create a session JWT.


        Decodes and verifies the HMAC-signed token, looks up or creates

        the user, and returns a JWT access token.'
      operationId: verify_magic_link_auth_magic_link_verify_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MagicLinkVerifyRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MagicLinkVerifyResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /auth/token:
    post:
      tags:
      - Auth
      summary: Create Api Token
      description: 'Create a long-lived API token for CLI/automation use.


        Requires an existing session (cookie or Bearer token). Users must

        authenticate first via OAuth or magic link, then create API tokens.


        The raw JWT is returned only once in the response. The caller must

        store it securely. The token includes a ``token_use: api`` claim to

        distinguish it from session JWTs.'
      operationId: create_api_token_auth_token_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenCreateRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenCreateResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /mcp/authorize/resume:
    get:
      tags:
      - Auth
      summary: Authorize Resume
      description: 'Resume authorization flow after login.


        The user was redirected to login, and after successful login they

        are sent back here. We now show the consent screen.'
      operationId: authorize_resume_mcp_authorize_resume_get
      parameters:
      - name: rid
        in: query
        required: true
        schema:
          type: string
          title: Rid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    AllSettingsResponse:
      properties:
        data_dir:
          type: string
          title: Data Dir
        gateway_port:
          type: integer
          title: Gateway Port
        llm:
          $ref: '#/components/schemas/LLMSettingsResponse'
        sync:
          $ref: '#/components/schemas/SyncSettingsResponse'
        deployment_mode:
          type: string
          title: Deployment Mode
          default: self_hosted
        user_plan:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: User Plan
      type: object
      required:
      - data_dir
      - gateway_port
      - llm
      - sync
      title: AllSettingsResponse
      description: Full application settings response.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    LLMSettingsResponse:
      properties:
        default_provider:
          type: string
          title: Default Provider
        fallback_enabled:
          type: boolean
          title: Fallback Enabled
        monthly_budget_usd:
          type: number
          title: Monthly Budget Usd
        providers:
          additionalProperties:
            $ref: '#/components/schemas/ProviderDetail'
          type: object
          title: Providers
      type: object
      required:
      - default_provider
      - fallback_enabled
      - monthly_budget_usd
      - providers
      title: LLMSettingsResponse
      description: LLM configuration section.
...
    MagicLinkRequest:
      properties:
        email:
          type: string
          title: Email
      type: object
      required:
      - email
      title: MagicLinkRequest
      description: Request to send a magic link login email.
    MagicLinkResponse:
      properties:
        status:
          type: string
          title: Status
        message:
          type: string
          title: Message
        dev_token:
          anyOf:
          - type: string
          - type: 'null'
          title: Dev Token
      type: object
      required:
      - status
      - message
      title: MagicLinkResponse
      description: Response after requesting a magic link.
    MagicLinkVerifyRequest:
      properties:
        token:
          type: string
          title: Token
      type: object
      required:
      - token
      title: MagicLinkVerifyRequest
      description: Request to verify a magic link token.
    MagicLinkVerifyResponse:
      properties:
        access_token:
          type: string
          title: Access Token
        token_type:
          type: string
          title: Token Type
          default: bearer
        user:
          additionalProperties: true
          type: object
          title: User
      type: object
      required:
      - access_token
      - user
      title: MagicLinkVerifyResponse
      description: Response after successfully verifying a magic link token.
...
    SettingsUpdateRequest:
      properties:
        monthly_budget_usd:
          anyOf:
          - type: number
          - type: 'null'
          title: Monthly Budget Usd
        default_provider:
          anyOf:
          - type: string
          - type: 'null'
          title: Default Provider
        fallback_enabled:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Fallback Enabled
        openai_compatible_base_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Openai Compatible Base Url
        openai_compatible_model:
          anyOf:
          - type: string
          - type: 'null'
          title: Openai Compatible Model
        openai_compatible_supports_json_response_format:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Openai Compatible Supports Json Response Format
        openai_compatible_supports_stream_usage:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Openai Compatible Supports Stream Usage
        openai_compatible_supports_reasoning_effort:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Openai Compatible Supports Reasoning Effort
        openai_compatible_max_tokens_field:
          anyOf:
          - type: string
          - type: 'null'
          title: Openai Compatible Max Tokens Field
        openai_compatible_reasoning_format:
          anyOf:
          - type: string
          - type: 'null'
          title: Openai Compatible Reasoning Format
      type: object
      title: SettingsUpdateRequest
      description: Request to update settings.
    SettingsUpdateResponse:
      properties:
        status:
          type: string
          title: Status
      type: object
      required:
      - status
      title: SettingsUpdateResponse
      description: Response after updating settings.
...
    SyncSettingsResponse:
      properties:
        enabled:
          type: boolean
          title: Enabled
        interval_minutes:
          type: integer
          title: Interval Minutes
        bucket:
          anyOf:
          - type: string
          - type: 'null'
          title: Bucket
      type: object
      required:
      - enabled
      - interval_minutes
      - bucket
      title: SyncSettingsResponse
      description: Cloud sync configuration section.
...
    TokenCreateRequest:
      properties:
        name:
          type: string
          title: Name
        expires_in_days:
          type: integer
          maximum: 365.0
          minimum: 1.0
          title: Expires In Days
          default: 30
      type: object
      required:
      - name
      title: TokenCreateRequest
      description: Request to create a long-lived API token.
    TokenCreateResponse:
      properties:
        access_token:
          type: string
          title: Access Token
        token_type:
          type: string
          title: Token Type
          default: bearer
        expires_at:
          type: string
          title: Expires At
        name:
          type: string
          title: Name
      type: object
      required:
      - access_token
      - expires_at
      - name
      title: TokenCreateResponse
      description: Response after creating a long-lived API token (shown only once).
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...